Do not require deskew and use MIP scaling
Originally the pipeline set deskew to TRUE even if it was not in
the config file if stage scanning was detected. To allow for more
flexiblity in the processing and to ensure that the config file is
a reflection of what will be enabled, deskewing is now optional
for stage scanning. A warning will be displayed if deskew is not
enabled for stage scanned experiments.

The xy pixel size was being used inconsistently for MIP generation.
The pipeline now explictly adds both the xy and z scaling to MIPs.

The combined effect of the above two changes was also to add an
option to scale the MIPs for stage scanning without deskew. The
options for MIPs in llsm-pipeline were updated to allow all the
options that are already available in mosaic-pipeline.