Add subpixel centering and optional z-resampling
Now using the "3D Objects Counter" plugin for identifying the the
peak of the bead. Centering is accomplished using Multiview
Reconstruction (and thus you need to install BigStitcher).

For deconvolving before deskewing, it is better to resample the PSF
in z beforehand rather than using the built in resampling. This is now
an option on the macro display (as a checkbox) and also requires
installation of BigStitcher.