Add option to deconvolve before skewing
Both the llsm-pipeline and mosaic-pipeline are updated to take an
option "decon-first" from the configuration file. In this case,
deconvolution is performed on the raw input images (or crop or
flatfield as appropriate), and then these files are deskewed. The
output of "decon-first" is saved in folders called
"decon_before_deskew" and "deskew_after_decon", and MIPs are created
for the "deskew_after_decon" files.

This addition to the pipeline requres a resampled PSF, which can
be created with center-psf.ijm. This macro has been updated to
clarify the dialog box and to add an option that does not require
a microscope angle. Note that the pipeline usage of decon-first
assumes that the input PSF has the same z-step size as the input
images, regardless of what is listed in the settings file.