Updating BDV naming convention for Camera B
The previous naming convention counted the channels taken on Camera A,
then added that value to all channels on Camera B. This works for
simpler acquisitions, but it is possible to have more complex
acquisitions. For example, channels could be acquired as CamB_ch0,
CamA_ch1, CamB_ch2. In this case, the prior naming scheme will end up
with both CamB_ch0 and CamA_ch1 as ch1. However, some shift in channel
names is necessary, as CamA_ch0 with CamB_ch0 is also a valid acquisition.

The new naming convention adds 10 to all Camera B channels. Thus
CamB_ch0 becomes ch10 and CamB_ch1 becomes ch11. All channels are also
now padded with zeros to two digits because of this change.

The LLSM pipeline does not have the same convention because the channel
names should be unique across channels. The llsm-pipeline file was
cleaned up to reflect this.