Add optional overwrite parameter to config
Default behavior of the pipeline is to re-process an entire folder,
regardless of exisiting tif files (equivalen to overwrite = true).
When overwrite = false, commands will only be generated for missing
files. The pipeline will identify missing files individually for each
tif. For example, if the flatfield file exists already, but the deskew
does not, only the deskew command will be generated.