Exclude zarr folders from parsing
Previously os.walk used topdown=False to look inside the subfolders of
previously processed folders. This would allow subexperiments to be
processed even if the parent folder was already processed. This was a
niche case, and topdown=False adds non-trivial time to the parsing, so
topdown has now been set to True.

In addition, any folders that contain .n5 or .zarr are excluded from the
search for settings.txt files. There is no reason for these folders to
contain the settings.txt file, and their parsing adds an infeasible
amount of time to the search for the settings files.